sq-poller: fix connection if the initial discovery failed
If the first attempt of discovery failed, the poller is not able to
connect to the node, even if it is up again. That's because the
init_ssh() function calls _fetch_init_dev_data() which is not
implemented in the Node object. This commit fixes this problem
alloing init_ssh() to call _fetch_init_dev_data() only when the devtype
is set.

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>