ospf: redo peer discovery, fix assert
In the presence of port channels and Vlans, we need to redo both peer
discovery and assert for OSPF. We first attempt to resolve peers via
lldp and failing that, we resolve peers via matching IPs. We ran into
some issues with unnumbered interfaces if we just use IPs to match.

We also improve the assert message if the issue is a loopback interface
is not marked as passive

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>

ospf assert: Fix error when loopback if !passive

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>