Static Chunker: Handle command limiting
Due to the inability of backend TACACS server to cope with high rates of
authentication, we need to add a method to rate limit the number of
simultaneous commands sent at any given time.

In this patch, the static chunker recognizes this parameter from the
config file, ensures its a multiple of the number of workers, and passes
the appropriately sliced count to each worker.

We also fix the controller to not terminate all workers on exception
in run_once mode.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>