Worker: Handle limiting simultaneous commands sent by a worker
To handle backend authentication servers that cannot sustain more than a
small number of authentication requests at a time, we add a parameter to
limit this number. This patch handles the worker end of implementing
such a limiter.

We use an asyncio counting semaphore to implement this limit. Every
device class uses this semaphore before establishing an SSH Session, and
before issuing a command.

This patch ensures that we interpret the parameter provided by the
controller to create an appropriately counted semaphore and pass it to
each device instance.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>