Node: Implement limiting the max simultaneous cmds sent
To handle backend authentication servers that cannot sustain more than a
small number of authentication requests at a time, this patch uses the
counting sempahore passed from the worker to every node instance thats
part of this worker to implement this limiter.

We define a custom context manager called limit_pipeline that ensures
that we acquire the sempahore before either attempting an SSH connection
or running a command.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>