Fix node discovery after first attempt (#723)
* sq-poller: fix discovery attempts

During the node initialization, if we are not able to discover the node
type, we reschedule the discovery (i.e. unable to connect). However,
if we fail the discovery again we do not update the next attempt time,
so from this moment, for each new command issued on the node we perform
an attempt. This commit fixes this problem updating the time for each
attempt.

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>

* sq-poller: fix connection if the initial discovery failed

If the first attempt of discovery failed, the poller is not able to
connect to the node, even if it is up again. That's because the
init_ssh() function calls _fetch_init_dev_data() which is not
implemented in the Node object. This commit fixes this problem
alloing init_ssh() to call _fetch_init_dev_data() only when the devtype
is set.

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>

* sq-poller: use lock to perform node discovery

As for each node we batch the commands we issue, we risk that multiple
tasks try to perform the node discovery. Use a lock so that only one among
those tasks is able to perform the discovery.

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>