Restore network commands for backward compatibility (#728)
* restore network sqobject functions

This commit adds back show, summarize, unique and top functions to
network sqobject. What it actually does, is to create a new namespace
object and call the correct function on that object.
An new argument called 'ignore_warning' can be passed to the sqobject
functions. If it's False (default) a warning is displayed showing, if
one, the correct function to call.

Signed-off-by: LucaNicosia <luca.nicosia@stardustsystems.net>

* add deprecated network commands

Signed-off-by: LucaNicosia <luca.nicosia@stardustsystems.net>

* add deprecated network routes

Signed-off-by: LucaNicosia <luca.nicosia@stardustsystems.net>

* tests

Signed-off-by: LucaNicosia <luca.nicosia@stardustsystems.net>

* fix typo

Signed-off-by: LucaNicosia <luca.nicosia@stardustsystems.net>