Netbox multi tag
Now it's possible to specify more than one tag.
All devices matching at least one of the tags in the list will be
picked.

If a tag is specified like 'tag1, tag2', a device in order to be picked
must be assigned to both tags.

The string format for tag is marked as deprecated.

Signed-off-by: LucaNicosia <luca.nicosia@stardustsystems.net>