Fix mackey to handle router MAC address
When we made a fix to add a "router" flag--applies only to nxos--we also
caused it to create what seemed like duplicate MAC table entries i.e.
one where the key fields are the same. In the case of NXOS, we have two
MAC address/VLAN pair with the same value but that differ in their
outgoing interface in the case of an SVI. So, we end up showing only one
of the two entries instead of both. This shouldn't affect any of the
learned MAC table entries, but this does affect what one sees on a
device versus what one sees with SuzieQ. This patch addresses the issue
by changing the mackey field construction to be similar to other such
fields with flags such as permanent.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>