Updated handling of network find
When using network find to find all the different IP/MAC bindings over
time, a bug to avoid showing duplicate entries prevented correct
operation of network find when used with start and end times or with
view=all. This patch addresses that.

In the process of fixing that issue, I found a few other issues. If the
user supplies the address of an interface on a device we're polling, the
results returned were inconsistent. This patch addresses that as well.
Such addresses are now returned with a type=interface to indicate that
these addresses are interface addresses on a device we're polling.

Finally, the logic to decide if an address was attached via a routed
interface or a bridged was incorrect. That's also fixed.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>