Fix coalesced file selection with start-time
In order to select a file from the coalesced files,
we wanted the start-time in the time interval of the file.
However, that meant that if the start-time is before the
oldest file, this is not selected. This commit fixes it
by selecting a file if the start-time comes before the
end-time of the file.

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>