Analyzer: more flexible parsing of operators
Thanks to splitting the strings on whitespace, if the user provided a
string such as "> 1500", it got split into two elements-- >, 1500 -- and
the parser failed to execute the appropriate query. Instead we required
the user to specify this as ">1500". Given that the error got caught in
a different place and the error message was unhelpful as to the reason,
this patch addresses the issue by parsing strings like "> 1500"
correctly.

The change is done in basicobj.py, but multiple routines had to be
changed:

* basicobj.py: validating input values returns an updated kwargs dict with
  the correct parsing of strings.
* routes/macs/arpnd/address: modified because they need to return the
  updated kwargs set
* lldp.py: needs to send ifindices, which are ints, as strings to avoid
  tripping up the code in basicobj.py. This was one of the few places
  where we were passing ints directly.
* parquetdb.py: Needed to strip the leading whitespace from "> 1400"
  when splitting it into the operator ">" and the value "1400".
  Furthermore if the user specifies both > and < in the query string,
  we treat the query as an AND and treat as a between the provided
  values.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>