set correctly namespaces list
In case of a filter on the topology summarize (f.e. 'topology summarize
vrf='intenet-vrf'), the variable self.nsns containing the list of
namespaces is populated using the unfiltered interface table.
On the other hands, the variable self.graphs is populated using the
namespaces contained in self.lsdb (the result of topology.get) that
may not contain one (or more) namespaces due to the filters.
During the computation of the summarize, the varialble self.graphs
is iterated using self.nsns entries. In some cases (like the example
given above) the code tries to access an unexistent key of self.graphs
causing the code to crash.
Now, self.nsns is populated using the data present in self.lsdb

Signed-off-by: LucaNicosia <luca.nicosia@stardustsystems.net>