Ensure consistent hostname setting
During the initial node discovery, we retrieve the hostname for a device
and use that. However, on initializing the SSH for the first time, we
also reinit the hostname, and sometimes from a different command than
whats present in the initial node discovery phase. The hostname needs to
be whatever is advertised in LLDP because that is how we find peers of
an interface and is essential for many tasks that we do.

This diff in hostname computed during node discovery from the one in
polling can cause us to recreate incorrect hostnames. This patch ensures
that we only ever use the hostname computed during ssh setup after node
discovery.

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>