sq-coalescer: fix multiple writes on the same parquet file
The patch fixes a bug causing the coalescer to write multiple times the same
coalesced file. This was because the coalescer wrote in advance the files
for all the coalescing windows detected during the sqpoller coalescing, and then
wrote again the file with the diff coming from the window. Additionally, now
the coalescer doesn't write anymore a coalesced file for each coalesing window
even when there aren't changes, as in the reading phase the engine is able to
handle missing coalescing periods, always retrieving the data according to
start and end time

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>