LLDP: Handle peerhost with fqdn but hostname without
In case of some NOS, the hostname from show version (or other commands)
doesn't contain the fully qualified domain name in the hostname while
the advertisement in LLDP does. This causes us to be unable to identify
peers which is necessary for much of the smarts in SuzieQ.

Instead of providing an additional knob to the user to ignore FQDN, we
can be smart in how we handle this. We pull up the list of peer hostname
in LLDP advertisements, map it with the available hostnames, and replace
the peer hostnames that have a hostname entry without FQDN to be without
FQDN when we return the LLDP results

Signed-off-by: Dinesh Dutt <dd.ps4u@gmail.com>
Signed-off-by: Claudio Lorina <claudio.lorina@stardustsystems.net>