sq-poller: add connection attempts control and `slow_host` for everybody
this patch improve the node connection handling with
a centralized _init_ssh() function, doing a 4 attempts.
If the all the attempts fail the node goes to a backoff
status, as there is no point in going on retrying,  and
it waits for some time before doing other attempts. In
the meanwhile all the issued commands on the node will fail.
This fixes some infinite loops in ios, iosxe, iosxr when the
reason was different than PermissionDenied.
Moreover the poller now control the number of connection attempts
avoiding too much too quick connection attempts.
This patch also allows to specify the `slow_host` configuration
for all the devtypes.

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>