sq-poller: fix issues with device connection (#799)
* sq-poller: fix error response if connection fails ios(xe)

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>

* sq-poller: avoid connection and disconnection if devtype is provided

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>

* sq-poller: Wait some time between first and second connection (ios)

with ios, iosxe and iosxr we need to close the connection
used for autodiscovery and to create a new persistent connection.
However, if connection after disconnection is done to quickly
authentication fails. To solve this problem we added a delay
to slow down the following connection.

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>

* sq-poller: increase connection backoff time with slow_host

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>

* sq-poller: improve connection log messages

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>

* sq-poller: do not call fetch_init_devdata if not connected

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>

* sq-poller: fix deadlock in ios(xe) iosxr init_ssh

ios(xe) and iosxr function called the _fetch_init_dev_data()
function without unlocking ssh_ready mutex. This might cause
a deadlock if init_ssh is called again.

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>

* sq-poller: fix call loop with _fetch_init_dev_data()

When we called the _fetch_init_dev_data() function and
we didn't have an active connection, the _init_ssh() fn
was called, that called again the _fetch_init_dev_data().
This caused a call loop that made the application crash for
too many recursive calls.

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>

* sq-poller: do not wait backoff time if retry=0 (iosxr)

With iosxr we waited the backoff time even when the retry
counter reached 0. This patch break the loop before the
wait time.

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>

* sq-poller: add connection attempts control and `slow_host` for everybody

this patch improve the node connection handling with
a centralized _init_ssh() function, doing a 4 attempts.
If the all the attempts fail the node goes to a backoff
status, as there is no point in going on retrying,  and
it waits for some time before doing other attempts. In
the meanwhile all the issued commands on the node will fail.
This fixes some infinite loops in ios, iosxe, iosxr when the
reason was different than PermissionDenied.
Moreover the poller now control the number of connection attempts
avoiding too much too quick connection attempts.
This patch also allows to specify the `slow_host` configuration
for all the devtypes.

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>