sq-poller: catch exception in node to allow service rescheduling
When the _exec_service() function raised an exception this was not
catched, causing the service to never receive a response and so
the service never reschedule itself. This PR fixes this issue
creating an error response when an exception is raised. Or reraising
the exception if the error is not recoverable.

Signed-off-by: Claudio Lorina <claudio.lorina@gmail.com>