controller plugins: deepcopy configuration
it's possible that the same configuration is used in multiple namespaces.
If this happens, it's possible that the configuration is modified during
the initialization of the plugin (see base_source contructor) causing
the next plugins to fail.
In order to fix this issue, pass to the deepcopy the object constructors

Signed-off-by: LucaNicosia <luca.nicosia@stardustsystems.net>