service: write deletions with unique timestamp
If the same uncaolesced file has more than one timestamp in the same
file we might have issues when coalesced and uncaolesced data coexists
in the data-directory.

In order to avoid the problem above, we must force the timestamp of all
records written to be the same. Since all the records to write will
have the same timestamp, we are going to set that timestamp also for the
deletions. If instead no data was returned from the device (f.e. the
device was unreachable) we use the current timestamp for the deleted
records.

Signed-off-by: LucaNicosia <luca.nicosia@stardustsystems.net>
Co-authored-by: Claudio Usai <claudio.usai@stardustsystems.net>